+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -6,6 +6,7 @@ const DatabaseTableName = {
 	MIGRATIONS: "migrations",
 	NOTIFICATIONS: "notifications",
 	PERMISSIONS: "permissions",
+	PLANNED_PATHS: "planned_paths",
 	POINTS_OF_INTEREST: "points_of_interest",
 	REVIEWS: "reviews",
 	ROUTE_CATEGORIES: "route_categories",