+++ apps/backend/src/libs/types/transaction-option.type.ts
@@ -1,0 +1,7 @@
+import { type Knex } from "knex";
+
+type TransactionOptions = {
+	transaction?: Knex.Transaction;
+};
+
+export { type TransactionOptions };