+++ apps/backend/src/modules/routes/route.entity.ts
@@ -1,7 +1,11 @@
-import { type Entity } from "~/libs/types/types.js";
+import { type Entity, type LineStringGeometry } from "~/libs/types/types.js";
 
 class RouteEntity implements Entity {
+	private createdByUserId: number;
 	private description: null | string;
+	private distance: number;
+	private duration: number;
+	private geometry: LineStringGeometry;
 	private id: null | number;
 	private name: string;
 	private pois: {
@@ -10,12 +14,20 @@ class RouteEntity implements Entity {
 	}[];
 
 	private constructor({
+		createdByUserId,
 		description,
+		distance,
+		duration,
+		geometry,
 		id,
 		name,
 		pois,
 	}: {
+		createdByUserId: number;
 		description: null | string;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		id: null | number;
 		name: string;
 		pois: {
@@ -24,13 +36,21 @@ class RouteEntity implements Entity {
 		}[];
 	}) {
 		this.id = id;
+		this.distance = distance;
+		this.duration = duration;
+		this.geometry = geometry;
 		this.name = name;
 		this.description = description;
 		this.pois = pois;
+		this.createdByUserId = createdByUserId;
 	}
 
 	public static initialize(data: {
+		createdByUserId: number;
 		description: string;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		id: number;
 		name: string;
 		pois: {
@@ -39,7 +59,11 @@ class RouteEntity implements Entity {
 		}[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdByUserId: data.createdByUserId,
 			description: data.description,
+			distance: data.distance,
+			duration: data.duration,
+			geometry: data.geometry,
 			id: data.id,
 			name: data.name,
 			pois: data.pois,
@@ -47,10 +71,18 @@ class RouteEntity implements Entity {
 	}
 
 	public static initializeList({
+		createdByUserId,
+		distance,
+		duration,
+		geometry,
 		id,
 		name,
 		pois,
 	}: {
+		createdByUserId: number;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		id: number;
 		name: string;
 		pois: {
@@ -59,7 +91,11 @@ class RouteEntity implements Entity {
 		}[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdByUserId,
 			description: null,
+			distance,
+			duration,
+			geometry,
 			id,
 			name,
 			pois,
@@ -67,11 +103,19 @@ class RouteEntity implements Entity {
 	}
 
 	public static initializeNew({
+		createdByUserId,
 		description,
+		distance,
+		duration,
+		geometry,
 		name,
 		pois,
 	}: {
+		createdByUserId: number;
 		description: string;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		name: string;
 		pois: {
 			id: number;
@@ -79,7 +123,11 @@ class RouteEntity implements Entity {
 		}[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdByUserId,
 			description,
+			distance,
+			duration,
+			geometry,
 			id: null,
 			name,
 			pois,
@@ -87,6 +135,10 @@ class RouteEntity implements Entity {
 	}
 
 	public toListObject(): {
+		createdByUserId: number;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		id: number;
 		name: string;
 		pois: {
@@ -95,6 +147,10 @@ class RouteEntity implements Entity {
 		}[];
 	} {
 		return {
+			createdByUserId: this.createdByUserId,
+			distance: this.distance,
+			duration: this.duration,
+			geometry: this.geometry,
 			id: this.id as number,
 			name: this.name,
 			pois: this.pois,
@@ -102,7 +158,11 @@ class RouteEntity implements Entity {
 	}
 
 	public toNewObject(): {
+		createdByUserId: number;
 		description: string;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		name: string;
 		pois: {
 			id: number;
@@ -110,14 +170,22 @@ class RouteEntity implements Entity {
 		}[];
 	} {
 		return {
+			createdByUserId: this.createdByUserId,
 			description: this.description as string,
+			distance: this.distance,
+			duration: this.duration,
+			geometry: this.geometry,
 			name: this.name,
 			pois: this.pois,
 		};
 	}
 
 	public toObject(): {
+		createdByUserId: number;
 		description: string;
+		distance: number;
+		duration: number;
+		geometry: LineStringGeometry;
 		id: number;
 		name: string;
 		pois: {
@@ -126,7 +194,11 @@ class RouteEntity implements Entity {
 		}[];
 	} {
 		return {
+			createdByUserId: this.createdByUserId,
 			description: this.description as string,
+			distance: this.distance,
+			duration: this.duration,
+			geometry: this.geometry,
 			id: this.id as number,
 			name: this.name,
 			pois: this.pois,