+++ apps/backend/src/modules/routes/route.model.ts
@@ -1,6 +1,7 @@
 import { Model } from "objection";
 
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
+import { type LineStringGeometry } from "~/libs/types/types.js";
 
 import { CategoryModel } from "../categories/category.model.js";
 import { PointsOfInterestModel } from "../points-of-interest/points-of-interest.model.js";
@@ -38,7 +39,11 @@ class RouteModel extends Model {
 		return DatabaseTableName.ROUTES;
 	}
 
+	public createdByUserId!: number;
 	public description!: string;
+	public distance!: number;
+	public duration!: number;
+	public geometry!: LineStringGeometry;
 	public id!: number;
 	public name!: string;
 	public pois!: {