+++ packages/shared/src/libs/types/line-string-geometry.type.ts
@@ -1,8 +1,9 @@
-import { type PointGeometry } from "./location-geometry.type.js";
+import { type LocationType } from "../enums/enums.js";
+import { type Coordinates } from "./coordinates.type.js";
 
 type LineStringGeometry = {
-	coordinates: PointGeometry["coordinates"][];
-	type: "LineString";
+	coordinates: Coordinates[];
+	type: typeof LocationType.LINE_STRING;
 };
 
 export { type LineStringGeometry };